Repair multiple functions.

+++ src/alg/multi_label_graphcut.hpp
@@ -32,7 +32,7 @@ struct ForSmoothFn
 		int dummyLabel;
 };
 
-inline int label_dist(int value);
+inline float label_dist(int value);
 float smoothFn(int p1, int p2, int l1, int l2, void *data);
 
 } /* namespace elib */
